chore(1-3293): add new query for daily data that uses  date ranges (#9150)
Add new methods to the store behind data usage metrics that accept date
ranges instead of a single month. The old data collection methods
re-route to the new ones instead, so the new methods are tested
implicitly.

Also deprecates the new endpoint that's not in use anywhere except in an
unused service method in Enterprise yet.

## Discussion point:

Accepts from and to params as dates for type safety. You can send unparseable strings, but if you send a date object, you know it'll work.

Leaves the use of the old method in `src/lib/features/instance-stats/instance-stats-service.ts` to keep changes small.