fix: dynamic highlight width
This makes the width of the highlight bars in the network dynamic and
based on the number of labels included in the chart.

Since the number of labels should always correspond to the number of
data points, this seems like a sensible approach.

With this, the label width will now be calculated on the fly, so even
if you resize the window or change the number of labels, the
highlighting will still work as expected.