chore: tests are timing out (#9264)
## About the changes
I figured that many of our tests are timing out and we can solve this by
increasing test timeout.

This is a signal that either our tests became slower or that our query
performance got worse. On the first point it might be due to having more
migrations than before and requiring more time. The second point is
harder to validate but something to keep an eye on