chore: add file and component names to `styled` output class names
Updates the `vite.config.mts` file to include config for emotion,
telling it to add `<file name>--<component name>` to the class names
of generated components.

It looks like this file is only used for development, but we should
double check that before merging. It .. Might also require some
snapshot updates.